+++ src/continuity/operators/deeponet.py
@@ -5,6 +5,7 @@ The DeepONet architecture.
 """
 
 import torch
+from typing import Optional
 from continuity.operators import Operator
 from continuity.operators.common import DeepResidualNetwork
 from continuity.operators.shape import OperatorShapes
@@ -28,6 +29,7 @@ class DeepONet(Operator):
         trunk_width: Width of trunk network
         trunk_depth: Depth of trunk network
         basis_functions: Number of basis functions
+        device: Device.
     """
 
     def __init__(
@@ -38,10 +40,9 @@ class DeepONet(Operator):
         trunk_width: int = 32,
         trunk_depth: int = 3,
         basis_functions: int = 8,
+        device: Optional[torch.device] = None,
     ):
-        super().__init__()
-
-        self.shapes = shapes
+        super().__init__(shapes, device)
 
         self.basis_functions = basis_functions
         self.dot_dim = shapes.v.dim * basis_functions
@@ -51,6 +52,7 @@ class DeepONet(Operator):
             output_size=self.dot_dim,
             width=trunk_width,
             depth=trunk_depth,
+            device=device,
         )
         # branch network
         branch_input_dim = shapes.u.num * shapes.u.dim
@@ -59,6 +61,7 @@ class DeepONet(Operator):
             output_size=self.dot_dim,
             width=branch_width,
             depth=branch_depth,
+            device=device,
         )
 
     def forward(