+++ src/continuity/operators/fno.py
@@ -22,6 +22,7 @@ class FourierNeuralOperator(NeuralOperator):
         depth: Number of Fourier layers.
         width: Latent dimension of the Fourier layers.
         act: Activation function. Default is tanh.
+        device: Device.
     """
 
     def __init__(
@@ -30,6 +31,7 @@ class FourierNeuralOperator(NeuralOperator):
         depth: int = 3,
         width: int = 3,
         act: Optional[torch.nn.Module] = None,
+        device: Optional[torch.device] = None,
     ):
         latent_shapes = OperatorShapes(
             x=shapes.x,
@@ -44,8 +46,9 @@ class FourierNeuralOperator(NeuralOperator):
             v=TensorShape(shapes.v.num, width),
         )
 
-        layers = [FourierLayer(latent_shapes) for _ in range(depth - 1)]
-        layers.append(FourierLayer(output_shapes))
+        layers = [
+            FourierLayer(latent_shapes, device=device) for _ in range(depth - 1)
+        ] + [FourierLayer(output_shapes, device=device)]
         layers = torch.nn.ModuleList(layers)
 
-        super().__init__(shapes, layers, act)
+        super().__init__(shapes, layers, act, device)